Made GigMenu into Popover component and added a Position field
